model load hack

+++ awq/models/auto.py
@@ -2,6 +2,7 @@ import os
 from transformers import AutoConfig
 from awq.models import *
 from awq.models.base import BaseAWQForCausalLM
+import torch
 
 AWQ_CAUSAL_LM_MODEL_MAP = {
     "mpt": MptAWQForCausalLM,
@@ -27,11 +28,11 @@ class AutoAWQForCausalLM:
                                'AutoAWQForCausalLM.from_quantized or AutoAWQForCausalLM.from_pretrained')
     
     @classmethod
-    def from_pretrained(self, model_path, trust_remote_code=True, safetensors=False) -> BaseAWQForCausalLM:
+    def from_pretrained(self, model_path, torch_dtype = torch.float16, trust_remote_code=True, safetensors=False) -> BaseAWQForCausalLM:
         model_type = check_and_get_model_type(model_path, trust_remote_code)
 
         return AWQ_CAUSAL_LM_MODEL_MAP[model_type].from_pretrained(
-            model_path, model_type, trust_remote_code=trust_remote_code, safetensors=safetensors
+            model_path, model_type, torch_dtype=torch_dtype, trust_remote_code=trust_remote_code, safetensors=safetensors
         )
 
     @classmethod
